+++ install/schema-sync/config.js
@@ -5,7 +5,8 @@
  *  collectionName:
  * 		watch: array of events to watch for changes, eg. 'posts.items',
  * 		excludeFields: (optional) array of fields to exclude from the export,
- * 		linkedFields: (optional) array of fields to treat as many-to-one relationships for hierarchy, eg. ['parent'],
+ * 		groupBy: (optional) array of fields to group the exported data into multiple files, eg. ['collection'] (per collection)
+ * 		linkedFields: (optional) array of fields to treat as many-to-one relationships for hierarchy, eg. ['parent']
  * 		getKey: (optional) function to get the key for the item, defaults to primary key found on schema,
  * 		query: (optional) query to use when exporting the collection, valid options are: (limit=-1 | filter | sort)
  * 		prefix: (optional) prefix the exported json file with this string (useful for separating test data from production data)